refactor: 분리한 SectionService 통합
(cherry picked from commit 44bcaf3b322eef2ee846ae72928f4e1ba53887f8)
