+++ src/main/java/subway/domain/line/dto/ShortestPathResponse.java
@@ -1,0 +1,5 @@
+package subway.domain.line.dto;
+
+public class ShortestPathResponse {
+
+}
